Do source map generation inline with the output generation.
Also moves the parser to a separate module. This reduces the length
of index.js.rst significantly.

Processing the source map at the same time as the output generation
is simpler. There is no inputLineNumbers array or off-by-one
adjustments and the code is easier to understand without the second
pass.

Always generates source map in response. This is easier than testing
for the presence of the inputFilename parameter and having an
overloaded interface.